+++ pre/lib/x86_64/mem.asm
